Finalizado o repositorio e iniciado a criação dos testes unitarios.

+++ trunk/VisProj/src/main/java/br/ufg/inf/visproj/model/Versao.java
@@ -52,7 +52,7 @@ public class Versao {
         return resultadoDaEquacao;
     }
 
-    public void setResultadoDaEquacao(int resultadoDaEquacao) {
+    public void setResultadoDaEquacao(Float resultadoDaEquacao) {
         this.resultadoDaEquacao = resultadoDaEquacao;
     }
